Celsius, Fahrenheit conversion. Module, reading and writing to file.

+++ Lista_3/3_1_file_read.py
@@ -5,7 +5,7 @@ with open('celcjusz.txt', 'r') as t:
 fahrenheit = open("fahrenheit.txt", 'w')
 for i in read:
     l = []
-    d = module_3_1.cel_fahr(int(i))
+    d = module_3_1.cel_fahr(float(i))
     fahrenheit.write(str(d)+"\n")
 fahrenheit.close()
